extract parsed schema to feature
